+++ demos/demo/lib/view_model/multi_demo_view_model.dart
@@ -1,0 +1,17 @@
+import 'package:lazx/lazx.dart';
+
+class MultiDemoViewModel extends LazxViewModel {
+  LazxData<int> counter = LazxData<int>(0);
+  LazxData<bool> show = LazxData<bool>(false);
+
+  @override
+  List<LazxData> get props => [counter, show];
+
+  void increment() {
+    counter.push(counter.value + 1);
+  }
+
+  void toggleShow() {
+    show.push(!show.value);
+  }
+}