+++ demos/lazx-ideas/lib/common/theme.dart
@@ -2,10 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:lazx_idea/common/constants.dart';
 
-import 'constants.dart';
-
 final theme = ThemeData(
-  accentColor: kAccentColor,
+  primaryColor: kAccentColor,
   bottomSheetTheme:
       BottomSheetThemeData(backgroundColor: Colors.black.withOpacity(0)),
   inputDecorationTheme: InputDecorationTheme(